+++ lib/volt/page/page.rb
@@ -63,14 +63,15 @@ class Page
     # Initialize tasks so we can get the reload message
     self.tasks if Volt.env.development?
 
-    # TODORW:
-    # channel.on('reconnected') do
-    #   @page._reconnected.cur = true
-    #
-    #   `setTimeout(function() {`
-    #     @page._reconnected.cur = false
-    #   `}, 2000);`
-    # end
+    if Volt.client?
+      channel.on('reconnected') do
+        @page._reconnected = true
+
+        `setTimeout(function() {`
+          @page._reconnected = false
+        `}, 2000);`
+      end
+    end
   end
 
   def flash